% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfeher_okabeito.R
\name{lfeher_okabeito}
\alias{lfeher_okabeito}
\title{Hex codes for Okabe-Ito colors}
\usage{
lfeher_okabeito(base_color = "#000000")
}
\arguments{
\item{base_color}{base color. Defaults to black.}
}
\description{
Hex codes for Okabe-Ito colors.
}
\examples{
lfeher_okabeito()

}
\keyword{lfeher}
