% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfeher_theme.R
\name{lfeher_theme}
\alias{lfeher_theme}
\title{L Feher preferred ggplot theme}
\usage{
lfeher_theme(base_size = 10, base_color = "black")
}
\arguments{
\item{base_size}{base size. Defaults to 10.}

\item{base_color}{base color. Defaults to black.}
}
\description{
L Feher preferred ggplot theme.
}
\examples{
lfeher_theme()

}
\keyword{lfeher}
