% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okabeito.R
\name{okabeito}
\alias{okabeito}
\title{Hex codes for Okabe-Ito colors}
\usage{
okabeito(base_color = "#000000")
}
\arguments{
\item{base_color}{base color. Defaults to black.}
}
\description{
Hex codes for Okabe-Ito colors.
}
\examples{
lfeheR::okabeito()

}
\keyword{lfeher}
