% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\title{Preferred ggplot theme}
\usage{
theme(base_size = 10, base_color = "black")
}
\arguments{
\item{base_size}{base size. Defaults to 10.}

\item{base_color}{base color. Defaults to black.}
}
\description{
Preferred ggplot theme.
}
\examples{
lfeheR::theme()

}
\keyword{lfeher}
